progress with seg fault
finally some progress with segfault, was caused my multiple things, accessClientData was being used wrong and modifying a local copy instead of the 'global' one.
Also there was duplicate variables with and without '_' infront and sometimes they were being used in mixed way.
Modified rest of the program to work with accessClientData function and format since basically whole response was not restructured for that